+++ src/groups/unitary_group.jl
@@ -1,11 +1,11 @@
 """
-   UnitaryGroup{T}
+    UnitaryGroup{T}
 
 The special orthogonal group ``$(_math(:U))(n)`` is the Lie group consisting of the [`MatrixMultiplicationGroupOperation`](@ref) on the
 manifold of rotations [`UnitaryMatrices`](@extref `Manifolds.GeneralUnitaryMatrices`) with absolute value of the determinant equal to one.
 
 # Constructor
-   UnitaryGroup(n; kwargs...)
+    UnitaryGroup(n::Int; kwargs...)
 
 Generate unitary group ``$(_math(:U))(n)``.
 All keyword arguments in `kwargs...` are passed on to [`Rotations`](@extref `Manifolds.Rotations`) as well.
@@ -16,7 +16,7 @@ const UnitaryGroup{T} = LieGroup{
     Manifolds.UnitaryMatrices{T,ManifoldsBase.ℂ},
 }
 
-function UnitaryGroup(n; kwargs...)
+function UnitaryGroup(n::Int; kwargs...)
     U = Manifolds.GeneralUnitaryMatrices(
         n, ManifoldsBase.ℂ, Manifolds.AbsoluteDeterminantOneMatrices; kwargs...
     )
